Move odata4j specific classes into separate package
