+++ nuget-feed/src/jetbrains/buildServer/nuget/feed/server/index/impl/NuGetArtifactsMetadataProvider.java
@@ -22,7 +22,7 @@ import jetbrains.buildServer.nuget.common.FeedConstants;
 import jetbrains.buildServer.nuget.common.PackageLoadException;
 import jetbrains.buildServer.nuget.feedReader.NuGetPackageAttributes;
 import jetbrains.buildServer.nuget.feed.server.NuGetServerSettings;
-import jetbrains.buildServer.nuget.feed.server.javaFeed.cache.ResponseCacheReset;
+import jetbrains.buildServer.nuget.feed.server.cache.ResponseCacheReset;
 import jetbrains.buildServer.nuget.server.version.FrameworkConstraints;
 import jetbrains.buildServer.serverSide.SBuild;
 import jetbrains.buildServer.serverSide.SBuildType;
