+++ nuget-feed/src/jetbrains/buildServer/nuget/feed/server/odata4j/NuGetFeedInMemoryEdmGenerator.java
@@ -14,11 +14,13 @@
  * limitations under the License.
  */
 
-package jetbrains.buildServer.nuget.feed.server.javaFeed;
+package jetbrains.buildServer.nuget.feed.server.odata4j;
 
 import com.intellij.openapi.diagnostic.Logger;
-import jetbrains.buildServer.nuget.feed.server.javaFeed.functions.NuGetFeedFunction;
-import jetbrains.buildServer.nuget.feed.server.javaFeed.functions.NuGetFeedFunctions;
+import jetbrains.buildServer.nuget.feed.server.MetadataConstants;
+import jetbrains.buildServer.nuget.feed.server.NuGetAPIVersion;
+import jetbrains.buildServer.nuget.feed.server.odata4j.functions.NuGetFeedFunction;
+import jetbrains.buildServer.nuget.feed.server.odata4j.functions.NuGetFeedFunctions;
 import jetbrains.buildServer.util.CollectionsUtil;
 import jetbrains.buildServer.util.filters.Filter;
 import org.jetbrains.annotations.NotNull;
