+++ nuget-feed/src/jetbrains/buildServer/nuget/feed/server/odata4j/PackageEntityEx.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package jetbrains.buildServer.nuget.feed.server.javaFeed;
+package jetbrains.buildServer.nuget.feed.server.odata4j;
 
 import jetbrains.buildServer.nuget.feed.server.NuGetServerSettings;
 import jetbrains.buildServer.nuget.feed.server.index.NuGetIndexEntry;
-import jetbrains.buildServer.nuget.feed.server.javaFeed.entity.PackageEntityAdapter;
+import jetbrains.buildServer.nuget.feed.server.odata4j.entity.PackageEntityAdapter;
 import org.jetbrains.annotations.NotNull;
 import org.odata4j.core.OAtomStreamEntity;
