+++ nuget-feed/src/jetbrains/buildServer/nuget/feed/server/odata4j/PackagesEntitySet.java
@@ -14,12 +14,21 @@
  * limitations under the License.
  */
 
-package jetbrains.buildServer.nuget.feed.server.javaFeed.cache;
+package jetbrains.buildServer.nuget.feed.server.odata4j;
+
+import jetbrains.buildServer.nuget.feed.server.MetadataConstants;
+import org.jetbrains.annotations.NotNull;
+import org.odata4j.edm.EdmEntitySet;
 
 /**
- * Created by Eugene Petrenko (eugene.petrenko@gmail.com)
- * Date: 23.04.13 9:34
+ * @author Evgeniy.Koshkin
  */
-public interface ResponseCacheReset {
-  void resetCache();
+public class PackagesEntitySet {
+  @NotNull
+  private static final EdmEntitySet.Builder myPackagesEntitySetBuilder
+          = new EdmEntitySet.Builder().setName(MetadataConstants.ENTITY_SET_NAME);
+
+  public static EdmEntitySet.Builder getBuilder() {
+    return myPackagesEntitySetBuilder;
+  }
 }
