+++ nuget-feed/src/jetbrains/buildServer/nuget/feed/server/odata4j/entity/PackageEntityAdapter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package jetbrains.buildServer.nuget.feed.server.javaFeed.entity;
+package jetbrains.buildServer.nuget.feed.server.odata4j.entity;
 
 import org.jetbrains.annotations.NotNull;
 import org.joda.time.LocalDateTime;
