+++ nuget-tests/src/jetbrains/buildServer/nuget/tests/integration/feed/server/NuGetJavaFeedIntegrationTestBase.java
@@ -26,8 +26,8 @@ import jetbrains.buildServer.nuget.feed.server.index.impl.PackagesIndexImpl;
 import jetbrains.buildServer.nuget.feed.server.index.impl.SemanticVersionsComparators;
 import jetbrains.buildServer.nuget.feed.server.index.impl.transform.DownloadUrlComputationTransformation;
 import jetbrains.buildServer.nuget.feed.server.index.impl.transform.IsPrereleaseTransformation;
-import jetbrains.buildServer.nuget.feed.server.javaFeed.NuGetProducerHolder;
-import jetbrains.buildServer.nuget.feed.server.javaFeed.functions.NuGetFeedFunctions;
+import jetbrains.buildServer.nuget.feed.server.odata4j.NuGetProducerHolder;
+import jetbrains.buildServer.nuget.feed.server.odata4j.functions.NuGetFeedFunctions;
 import jetbrains.buildServer.nuget.tests.integration.Paths;
 import jetbrains.buildServer.serverSide.metadata.BuildMetadataEntry;
 import jetbrains.buildServer.serverSide.metadata.MetadataStorage;
