+++ nuget-tests/src/jetbrains/buildServer/nuget/tests/server/feed/server/PackageIndexTest.java
@@ -26,7 +26,7 @@ import jetbrains.buildServer.nuget.feed.server.index.impl.transform.AccessCheckT
 import jetbrains.buildServer.nuget.feed.server.index.impl.transform.DownloadUrlComputationTransformation;
 import jetbrains.buildServer.nuget.feed.server.index.impl.transform.IsPrereleaseTransformation;
 import jetbrains.buildServer.nuget.feed.server.index.impl.transform.SamePackagesFilterTransformation;
-import jetbrains.buildServer.nuget.feed.server.javaFeed.entity.PackageEntityAdapter;
+import jetbrains.buildServer.nuget.feed.server.odata4j.entity.PackageEntityAdapter;
 import jetbrains.buildServer.nuget.tests.integration.feed.server.MockExternalIdTransformation;
 import jetbrains.buildServer.serverSide.ProjectManager;
 import jetbrains.buildServer.serverSide.auth.AuthorityHolder;
